requirements.txt
numpy
opencv-python
pillow
easyocr
matplotlib

inference[transformers]
inference[sam]
inference[clip]
inference[gaze]
inference[grounding-dino]
inference[yolo-world]
inference

supervision